Added license. Public domain!
Considered using 0BSD, but… it irks me ever so slightly that it has “BSD” in the name despite being based on the ISC license, and more importantly, I *want* the code to be in the public domain for ideological reasons. Whether or not the end result would be identical in which rights users would be granted, utilizing this wonderful thing known as the “public domain” is something that I wish was more common (and I wish was easier – it pains me that copyright laws are written under the assumption that everyone *wants* copyright).