Switched from `.ogg` extension to `.opus` for ideological reasons.
Ideologically, I support the idea of an open container format – but as it turns out, Ogg is a *terrible* container format that's hindered its own adoption by attempting to be inadvisably generic. I was keeping the `.ogg` extension in order to venerate the veteran container format even now that Opus is on the scene – the codec has enough press, and I didn't want it to steal all the thunder from Ogg – but now that I have some more experience with the specifics of Ogg, I wish for all its thunder to be stolen. Opus deserves a better container format. And there *are* better open container formats! Such as Matroška (or is that Matroska?)! But I can't imagine many podcast apps support `.mka` files!

Either way, ideologically, using the `.opus` extension is… Well, I'm a bit torn, because on the one hand it draws attention away from Ogg (so that it may hopefully one day be superceded by a different open container format), but on the other, it means that `.opus` files will always be Ogg Opus. Although that latter boat has likely already sailed, so I suppose only the former is a factor in any real sense (and even then… the impact this one change is going to have is infinitesimal, hahah).